(refactor) app base theming
