+++ app/src/App.tsx
@@ -1,40 +1,17 @@
+import { configureStore } from '@reduxjs/toolkit'
 import React from 'react'
+import { Provider } from 'react-redux'
+import { ThemeProvider } from 'styled-components'
 import Web3Provider from 'web3-react'
-import { createGlobalStyle, ThemeProvider } from 'styled-components'
+
 import 'react-datepicker/dist/react-datepicker.css'
 import 'sanitize.css'
-import theme from './theme'
-import connectors from './util/connectors'
 import { Main } from './components'
-import { WalletConnectStylesOverride } from './common/wallet_connect_styles_override'
-import { configureStore } from '@reduxjs/toolkit'
 import balanceReducer from './store/reducer'
-import { Provider } from 'react-redux'
+import theme from './theme'
+import { GlobalStyle } from './theme/component_styles/global_style'
+import connectors from './util/connectors'
 
-type ThemeType = typeof theme
-
-const GlobalStyle = createGlobalStyle<{ theme: ThemeType }>`
-  body {
-    -moz-osx-font-smoothing: grayscale;
-    -webkit-font-smoothing: antialiased;
-    background-color: ${props => props.theme.colors.mainBodyBackground};
-    font-family: ${props => props.theme.fonts.fontFamily};
-    font-size: ${props => props.theme.fonts.defaultSize};
-  }
-
-  code {
-    font-family: ${props => props.theme.fonts.fontFamilyCode};
-  }
-
-  body,
-  html,
-  #root {
-    height: 100vh;
-  }
-
-  ${WalletConnectStylesOverride}
-
-`
 const store = configureStore({ reducer: balanceReducer })
 
 const App: React.FC = () => {
