+++ app/src/theme/component_styles/global_style.ts
@@ -1,0 +1,29 @@
+import { createGlobalStyle } from 'styled-components'
+
+import theme from '../index'
+import { WalletConnectStylesOverride } from '../overrides/wallet_connect_styles_override'
+
+type ThemeType = typeof theme
+
+export const GlobalStyle = createGlobalStyle<{ theme: ThemeType }>`
+  body {
+    -moz-osx-font-smoothing: grayscale;
+    -webkit-font-smoothing: antialiased;
+    background-color: ${props => props.theme.colors.mainBodyBackground};
+    font-family: ${props => props.theme.fonts.fontFamily};
+    font-size: ${props => props.theme.fonts.defaultSize};
+  }
+
+  code {
+    font-family: ${props => props.theme.fonts.fontFamilyCode};
+  }
+
+  body,
+  html,
+  #root {
+    height: 100vh;
+  }
+
+  ${WalletConnectStylesOverride}
+
+`
