+++ app/src/theme/overrides/wallet_connect_styles_override.tsx
